Dash added as well as ActionBuffer. Minor changes elsewhere (mainly Timer and Attack). Added new IDasher, which Agent implements.

+++ Capstone Project/GameObjects/Entities/Enemy.cs
@@ -5,7 +5,8 @@ namespace Capstone_Project.GameObjects.Entities
 {
     public class Enemy : Agent
     {
-        public Enemy(Subsprite subsprite, Vector2 position, int vitality, float damage, float attackRange = 100f, float defence = 0f, int size = 100, int speed = 100) : base(subsprite, position, vitality, damage, attackRange, defence, size, speed)
+        public Enemy(Subsprite subsprite, Vector2 position, int vitality, float damage, float attackRange = 100f, float defence = 0f, int size = 100, float speed = 100f)
+            : base(subsprite, position, vitality, damage, attackRange, defence, size, speed)
         {
             //Direction = new Vector2(0, 1);
         }
